% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{big_normal_logit_single_gibbs}
\alias{big_normal_logit_single_gibbs}
\title{Bayesian logistic regression with normal priors by single
variable update Gibbs sampler when predictor matrix is a big.matrix}
\usage{
big_normal_logit_single_gibbs(
  y,
  bigX,
  mcmc_sample = 500L,
  burnin = 500L,
  thinning = 1L,
  A_tau = 1,
  verbose = 0L
)
}
\arguments{
\item{y}{vector of n binary outcome variables taking values 0 or 1}

\item{mcmc_sample}{number of MCMC iterations saved}

\item{burnin}{number of iterations before start to save}

\item{thinning}{number of iterations to skip between two saved iterations}

\item{A_tau}{scale parameter in the half Cauchy prior of the ratio between the coefficient variance and the noise variance}

\item{X}{n x p sparse matrix of candidate predictors}
}
\value{
a list object consisting of three components
\describe{
\item{post_mean}{a list object of four components for posterior mean statistics}
\describe{
\item{betacoef}{a vector of posterior mean of p regression coeficients}
\item{tau2}{posterior mean of the ratio between prior regression coefficient variances and the noise variance}
\item{mu}{a vector of posterior predictive mean for linear predictor of the n training sample}
\item{prob}{a vector of posterior predictive probability of the n training sample}
}
\item{mcmc}{a list object of three components for MCMC samples}
\describe{
\item{betacoef}{a matrix of MCMC samples for p regression coeficients. Each column is one MCMC sample}
\item{tau2}{a vector of MCMC samples of global shrinkage parameters}
}
\item{elapsed}{running time}
}
}
\description{
Bayesian logistic regression with normal priors by single
variable update Gibbs sampler when predictor matrix is a big.matrix
}
\examples{
set.seed(2022)
library(bigmemory)
dat1 <- sim_logit_reg(n=2000,p=200,X_cor=0.9,X_var=10,q=10,beta_size=5)
dat1$X <- as.big.matrix(dat1$X)
res1 <- with(dat1,big_normal_logit_single_gibbs(y,X@address))
res1_glmnet <- with(dat1,wrap_glmnet(y,X,alpha=0.5,family=binomial()))
dat2 <- sim_logit_reg(n=200,p=2000,X_cor=0.9,X_var=10,q=10,beta_size=5)
res2 <- with(dat2,big_normal_logit_single_gibbs(y,X@address,burnin=5000))
res2_glmnet <- with(dat2,wrap_glmnet(y,X,alpha=0.5,family=binomial()))
tab <- data.frame(rbind(comp_sparse_SSE(dat1$betacoef,res1$post_mean$betacoef),
comp_sparse_SSE(dat1$betacoef,res1_glmnet$betacoef),
comp_sparse_SSE(dat2$betacoef,res2$post_mean$betacoef),
comp_sparse_SSE(dat2$betacoef,res2_glmnet$betacoef)),
time=c(res1$elapsed,res1_glmnet$elapsed,res2$elapsed,res2_glmnet$elapsed))
rownames(tab)<-c("n = 2000, p = 200 Bayes","n = 2000, p = 200 glmnet",
"n = 200, p = 2000 Bayes","n = 200, p = 2000 glmnet")
normal_logit_tab <- tab
print(normal_logit_tab)
}
\author{
Jian Kang <jiankang@umich.edu>
}
