% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBayesReg.R
\name{comp_class_acc}
\alias{comp_class_acc}
\title{Evaluate classification accuracy}
\usage{
comp_class_acc(pred, obs, levels = c(1, 0))
}
\arguments{
\item{pred}{a vector of predicted class indicators, taking two distict values, e.g. 0,1}

\item{obs}{a vector of observed class indicators, taking two distict values, e.g. 0,1}

\item{levels}{indicator class values, default c(1,0)}
}
\value{
a vector of five elements TPR, FPR, FDR, ACC, BA
\describe{
\item{TPR}{True positive rate}
\item{FPR}{False positive rate}
\item{FDR}{False discovery rate}
\item{ACC}{Accuracy}
\item{BA}{Balanced Accuracy}
}
}
\description{
Evaluate classification accuracy
}
\examples{
dat <- sim_logit_reg(n=2000,p=20,X_cor=0.9,q=6)
res <- with(dat,fast_normal_logit(y,X))
res_glmnet <- with(dat,wrap_glmnet(y,X,family=binomial()))
tab = rbind(comp_class_acc(as.numeric(res$post_mean$prob>0.5),dat$y),
comp_class_acc(as.numeric(predict(res_glmnet$glmnet_fit,dat$X,type = "response")>0.5),dat$y))
rownames(tab) = c("Bayes","glmnet")
print(tab)
}
\author{
Jian Kang <jiankang@umich.edu>
}
