% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBayesReg.R
\name{comp_sparse_SSE}
\alias{comp_sparse_SSE}
\title{Compute the sum of squared errors for sparse regression coefficients}
\usage{
comp_sparse_SSE(truebeta, estbeta)
}
\arguments{
\item{truebeta}{a vector of true regression coefficients}

\item{estbeta}{a vector of estimated regression coefficients}
}
\value{
a vector of three measures
\describe{
\item{overall}{Sum of squared errors for all regression coefficients}
\item{nonzero}{Sum of squared errors for nonzero regression coefficients}
\item{zero}{Sum of squared errors for zero regression coefficients}
}
}
\description{
Compute the sum of squared errors for sparse regression coefficients
}
\examples{
dat <- sim_linear_reg(n=2000,p=200,X_cor=0.9,q=6)
res <- with(dat1,fast_normal_lm(y,X))
print(comp_sparse_SSE(dat$betacoef,res$post_mean$betacoef))
}
\author{
Jian Kang <jiankang@umich.edu>
}
