% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_mfvb_normal_lm}
\alias{fast_mfvb_normal_lm}
\title{Fast Mean Field Varational Bayesian linear regression with normal priors}
\usage{
fast_mfvb_normal_lm(
  y,
  X,
  max_iter = 500L,
  a_sigma = 0.01,
  b_sigma = 0.01,
  A_tau = 1,
  tol = 1e-05,
  t_sigma2_eps_0 = 0,
  t_tau2_0 = 0
)
}
\arguments{
\item{y}{vector of n outcome variables}

\item{X}{n x p matrix of candidate predictors}

\item{max_iter}{max number of iterations}

\item{a_sigma}{shape parameter in the inverse gamma prior of the noise variance}

\item{b_sigma}{rate parameter in the inverse gamma prior of the noise variance}

\item{A_tau}{scale parameter in the half Cauchy prior of the ratio between the coefficient variance and the noise variance}
}
\value{
a list object consisting of two components
\describe{
\item{post_mean}{a list object of four components for posterior mean statistics}
\describe{
\item{mu}{a vector of posterior predictive mean of the n training sample}
\item{betacoef}{a vector of posterior mean of p regression coeficients}
\item{sigma2_eps}{posterior mean of the noise variance}
\item{tau2}{posterior mean of the ratio between prior regression coefficient variances and the noise variance}
}
\item{trace}{a list object for parameter updates}
\describe{
\item{mu}{a vector of posterior predictive mean of the n training sample}
\item{betacoef}{a vector of posterior mean of p regression coeficients}
\item{sigma2_eps}{posterior mean of the noise variance}
\item{tau2}{posterior mean of the ratio between prior regression coefficient variances and the noise variance}
}
}
}
\description{
Fast Mean Field Varational Bayesian linear regression with normal priors
}
\examples{
set.seed(2022)
dat1 <- sim_linear_reg(n=2000,p=200,X_cor=0.9,q=6)
res1 <- with(dat1,fast_mfvb_normal_lm(y,X))
dat2 <- sim_linear_reg(n=200,p=2000,X_cor=0.9,q=6)
res2 <- with(dat2,fast_mfvb_normal_lm(y,X))
tab <- data.frame(rbind(comp_sparse_SSE(dat1$betacoef,res1$post_mean$betacoef),
comp_sparse_SSE(dat2$betacoef,res2$post_mean$betacoef)),
time=c(res1$elapsed,res2$elapsed))
rownames(tab)<-c("n = 2000, p = 200","n = 200, p = 2000")
fast_normal_tab <- tab
print(fast_normal_tab)
}
\author{
Jian Kang <jiankang@umich.edu>
}
