% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_mfvb_normal_logit_single}
\alias{fast_mfvb_normal_logit_single}
\title{Fast single variable update mean field variational Bayesian
logistic regression with normal priors}
\usage{
fast_mfvb_normal_logit_single(
  y,
  X,
  max_iter = 5000L,
  tol = 1e-05,
  A = 10,
  in_E_inv_tau_sq = 1,
  in_E_omega = NULL,
  in_E_beta = NULL
)
}
\arguments{
\item{y}{vector of n binrary outcome variables taking values 0 or 1}

\item{X}{n x p matrix of candidate predictors}

\item{max_iter}{maximum number of iterations}

\item{tol}{the tolerance for the parameter changes}

\item{A}{scale parameter in the half Cauchy prior for regression coefficients}

\item{in_E_inv_tau_sq}{numeric scalar for the initial value for inverse of tau squared}

\item{in_E_omega}{numeric vector of length n for the initial values for all omega's}

\item{in_E_beta}{numeric vector of length p for the initial values for all beta's}
}
\value{
a list object consisting of three components
\describe{
\item{post_mean}{a list object of four components for posterior mean statistics}
\describe{
\item{betacoef}{a vector of posterior mean of p regression coeficients}
\item{inv_tau_sq}{posterior mean of inverse tau square}
\item{omega}{a vector of posterior mean of omega}
}
\item{trace}{a list object of two components for the trace of parameter updates}
\describe{
\item{inv_tau_sq}{a vector of trace for inverse of tau_sq}
\item{sum_beta_sq}{a vector of trace for inverse of summation of beta_sq}
}
\item{elapsed}{running time}
}
}
\description{
Fast single variable update mean field variational Bayesian
logistic regression with normal priors
}
\examples{
set.seed(2022)
dat1 <- sim_logit_reg(n=2000,p=20,X_cor=0.5,X_var=1,q=10,beta_size=5)
split <- train_test_splits(dat1$n)
dat1$train_idx <- split$train_idx
dat1$test_idx <- split$test_idx
res1 <- with(dat1,fast_mfvb_normal_logit_single(y[train_idx],X[train_idx,]))
res1_mcmc <- with(dat1,fast_normal_logit(y[train_idx],X[train_idx,]))
res1_glmnet <- with(dat1,wrap_glmnet(y[train_idx],X[train_idx,],alpha=0.5,family=binomial()))


tab <- data.frame(rbind(comp_sparse_SSE(dat1$betacoef,res1$post_mean$betacoef),
comp_sparse_SSE(dat1$betacoef,res1_glmnet$betacoef),
comp_sparse_SSE(dat1$betacoef,res1_mcmc$post_mean$betacoef)),
time=c(res1$elapsed,res1_glmnet$elapsed,
res1_mcmc$elapsed))

rownames(tab)<-c("n = 2000, p = 20 MFVB","n = 2000, p = 20 glmnet",
"n = 2000, p = 20 MCMC")
normal_logit_tab <- tab
print(normal_logit_tab)
}
\author{
Jian Kang <jiankang@umich.edu>
}
