% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_normal_multi_lm}
\alias{fast_normal_multi_lm}
\title{Fast Bayesian linear regression with normal priors with multiple outcome variables}
\usage{
fast_normal_multi_lm(
  y,
  X,
  mcmc_sample = 500L,
  burnin = 500L,
  thinning = 1L,
  a_sigma = 0.01,
  b_sigma = 0.01,
  A_tau = 10,
  mcmc_output = TRUE
)
}
\arguments{
\item{y}{n x q matrix of q outcome variables with n observations}

\item{X}{n x p matrix of p candidate predictors with n observations}

\item{mcmc_sample}{number of MCMC iterations saved}

\item{burnin}{number of iterations before start to save}

\item{thinning}{number of iterations to skip between two saved iterations}

\item{a_sigma}{shape parameter in the inverse gamma prior of the noise variance}

\item{b_sigma}{rate parameter in the inverse gamma prior of the noise variance}

\item{A_tau}{scale parameter in the half Cauchy prior of the ratio between the coefficient variance and the noise variance}
}
\value{
a list object consisting of two components
\describe{
\item{post_mean}{a list object of four components for posterior mean statistics}
\describe{
\item{mu}{a matrix of posterior predictive mean of the n training sample}
\item{betacoef}{a matrix of posterior mean of q x p regression coeficients}
\item{sigma2_eps}{posterior mean of the noise variance}
\item{tau2}{posterior mean of the ratio between prior regression coefficient variances and the noise variance}
}
\item{mcmc}{a list object of three components for MCMC samples}
\describe{
\item{mu}{an array of posterior predictive mean of the n training sample}
\item{betacoef}{an array of posterior mean of p by q regression coeficients}
\item{sigma2_eps}{posterior mean of the noise variance}
\item{tau2}{posterior mean of the ratio between prior regression coefficient variances and the noise variance}
}
}
}
\description{
Fast Bayesian linear regression with normal priors with multiple outcome variables
}
\examples{
set.seed(2024)
dat1 <- sim_linear_reg_multi(n=2000,p=200,X_cor=0.9,q=6)
res1 <- with(dat1,fast_normal_multi_lm(y,X,mcmc_output=FALSE))
dat2 <- sim_linear_reg(n=200,p=2000,X_cor=0.9,q=6)
res2 <- with(dat2,fast_normal_multi_lm(y,X))
tab <- data.frame(rbind(comp_sparse_SSE(dat1$betacoef,res1$post_mean$betacoef),
comp_sparse_SSE(dat2$betacoef,res2$post_mean$betacoef)),
time=c(res1$elapsed,res2$elapsed))
rownames(tab)<-c("n = 2000, p = 200","n = 200, p = 2000")
fast_normal_tab <- tab
print(fast_normal_tab)
}
\author{
Jian Kang <jiankang@umich.edu>
}
