% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log1mexpm}
\alias{log1mexpm}
\title{Accurately compute log(1-exp(-x)) for x > 0}
\usage{
log1mexpm(x)
}
\arguments{
\item{x}{a vector of nonnegative numbers}
}
\value{
a vector of values of log(1-exp(-x))
}
\description{
Accurately compute log(1-exp(-x)) for x > 0
}
\examples{
x <- seq(0,10,length=1000)
y <- log1mexpm(x)
plot(x,y,type="l")
}
\author{
Jian Kang <jiankang@umich.edu>
}
