% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log1pexp}
\alias{log1pexp}
\title{Accurately compute log(1+exp(x))}
\usage{
log1pexp(x)
}
\arguments{
\item{x}{a vector of real numbers}
}
\value{
a vector of values of log(1+exp(x))
}
\description{
Accurately compute log(1+exp(x))
}
\examples{
x <- seq(-100,100,length=1000)
y <- log1pexp(x)
plot(x,y,type="l")
}
\author{
Jian Kang <jiankang@umich.edu>
}
