% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_fast_logit}
\alias{predict_fast_logit}
\title{Prediction with fast Bayesian logistic regression fitting}
\usage{
predict_fast_logit(model_fit, X_test, alpha = 0.95, cutoff = 0.5)
}
\arguments{
\item{model_fit}{output list object of fast Bayesian logistic regression fitting (see value of \link{fast_horseshoe_lm} as an example)}

\item{X_test}{\eqn{n} by \eqn{p} matrix of predictors for the test data}

\item{alpha}{posterior predictive credible level \eqn{\alpha \in (0,1)}. The default value is \eqn{0.95}.}

\item{cutoff}{threshold value for posterior predicitve probablity. The default value is 0.5}
}
\value{
a list object consisting of three components
\describe{
\item{class}{a vector of \eqn{n} predicted class indicators}
\item{mean}{a vector of \eqn{n} posterior predictive mean probabilities}
\item{ucl}{a vector of \eqn{n}  posterior \eqn{\alpha} level upper credible limits of probabilities}
\item{lcl}{a vector of \eqn{n}  posterior \eqn{\alpha} level lower credible limits of probabilities}
\item{median}{a vector of \eqn{n}  posterior predictive median probabilities}
\item{sd}{a vector of \eqn{n}  standard deviations of posterior predictive probabilities}
}
}
\description{
Prediction with fast Bayesian logistic regression fitting
}
\examples{
dat <- sim_logit_reg(n=2000,p=200,X_cor=0.9,q=6)
train_idx = 1:round(length(dat$y)/2)
test_idx = setdiff(1:length(dat$y),train_idx)
res <- fast_normal_logit(dat$y[train_idx],dat$X[train_idx,])
pred_res <- predict_fast_logit(res,dat$X[test_idx,])
plot(dat$prob[test_idx,],pred_res$mean,
type="p",pch=19,cex=0.5,col="blue",asp=1,xlab="True Probabilities",
ylab = "Predicted Probabilities")
abline(0,1)
print(comp_class_acc(pred_res$class,dat$y[test_idx]))
}
\author{
Jian Kang <jiankang@umich.edu>
}
