% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_fast_mfvb_lm}
\alias{predict_fast_mfvb_lm}
\title{Prediction with fast mean field variational Bayesian linear regression fitting}
\usage{
predict_fast_mfvb_lm(model_fit, X_test)
}
\arguments{
\item{model_fit}{output list object of fast Bayesian linear regression fitting (see value of \link{fast_horseshoe_lm} as an example)}

\item{X_test}{\eqn{n} by \eqn{p} matrix of predictors for the test data}

\item{alpha}{posterior predictive credible level \eqn{\alpha \in (0,1)}. The default value is \eqn{0.95}.}
}
\value{
a list object consisting
\describe{
\item{mean}{a vector of \eqn{n} posterior predictive mean values}
}
}
\description{
Prediction with fast mean field variational Bayesian linear regression fitting
}
\examples{
dat <- sim_linear_reg(n=2000,p=200,X_cor=0.9,q=6)
train_idx = 1:round(length(dat$y)/2)
test_idx = setdiff(1:length(dat$y),train_idx)
res <- fast_mfvb_normal_lm(dat$y[train_idx],dat$X[train_idx,])
pred_res <- predict_fast_mfvb_lm(res,dat$X[test_idx,])
plot(dat$y[test_idx,],pred_res$mean,
type="p",pch=19,cex=0.5,col="blue",asp=1,xlab="Observations",
ylab = "Predictions")
abline(0,1)
}
\author{
Jian Kang <jiankang@umich.edu>
}
