% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_fast_mfvb_logit}
\alias{predict_fast_mfvb_logit}
\title{Prediction with fast mean field variational Bayesian logistic regression fitting}
\usage{
predict_fast_mfvb_logit(model_fit, X_test, alpha = 0.95, cutoff = 0.5)
}
\arguments{
\item{model_fit}{output list object of fast mean field variational Bayesian logistic regression fitting (see value of \link{fast_mfvb_normal_logit} as an example)}

\item{X_test}{\eqn{n} by \eqn{p} matrix of predictors for the test data}

\item{alpha}{posterior predictive credible level \eqn{\alpha \in (0,1)}. The default value is \eqn{0.95}.}

\item{cutoff}{threshold value for posterior predicitve probablity. The default value is 0.5}
}
\value{
a list object consisting of three components
\describe{
\item{class}{a vector of \eqn{n} predicted class indicators}
\item{prob}{a vector of \eqn{n} posterior predictive mean probabilities}
}
}
\description{
Prediction with fast mean field variational Bayesian logistic regression fitting
}
\examples{
dat <- sim_logit_reg(n=2000,p=200,X_cor=0.9,q=6,beta_size=5)
train_idx = 1:round(length(dat$y)/2)
test_idx = setdiff(1:length(dat$y),train_idx)
res_mcmc <- fast_normal_logit(dat$y[train_idx],dat$X[train_idx,])
res <- fast_mfvb_normal_logit(dat$y[train_idx],dat$X[train_idx,])
pred_res <- predict_fast_mfvb_logit(res,dat$X[test_idx,])
pred_res_mcmc <- predict_fast_logit(res_mcmc,dat$X[test_idx,])
par(mfrow=c(1,2))
plot(dat$prob[test_idx,],pred_res$prob,
type="p",pch=19,cex=0.5,col="blue",asp=1,xlab="True Probabilities",
ylab = "Predicted Probabilities",main="MFVB")
abline(0,1)
plot(dat$prob[test_idx,],pred_res_mcmc$mean,
type="p",pch=19,cex=0.5,col="blue",asp=1,xlab="True Probabilities",
ylab = "Predicted Probabilities",main="MCMC")
abline(0,1)
mfvb <- comp_class_acc(pred_res$class,dat$y[test_idx])
mcmc <- comp_class_acc(pred_res_mcmc$class,dat$y[test_idx])
tab <- rbind(mfvb,mcmc)
print(tab)
}
\author{
Jian Kang <jiankang@umich.edu>
}
