% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rand_left_trucnorm}
\alias{rand_left_trucnorm}
\title{Simulate left truncated normal distribution}
\usage{
rand_left_trucnorm(n, mu, sigma, lower, ratio = 1)
}
\arguments{
\item{n}{sample size}

\item{mu}{the location parameter}

\item{sigma}{the scale parameter}

\item{lower}{the lower bound}
}
\value{
a vector of random numbers from a left truncated
normal distribution
}
\description{
Simulate left truncated normal distribution
}
\examples{
r <- rand_left_trucnorm(1000, mu=100, sigma=0.1, lower=200)
hist(r)
}
\author{
Jian Kang <jiankang@umich.edu>
}
