% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rand_left_trucnorm0}
\alias{rand_left_trucnorm0}
\title{Simulate left standard truncated normal distribution}
\usage{
rand_left_trucnorm0(n, lower, ratio = 1)
}
\arguments{
\item{n}{sample size}

\item{lower}{the lower bound}
}
\value{
a vector of random numbers from a left standard truncated
normal distribution
}
\description{
Simulate left standard truncated normal distribution
}
\examples{
r0 <- rand_left_trucnorm0(1000, lower=2)
hist(r0)
}
\author{
Jian Kang <jiankang@umich.edu>
}
