% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rand_right_trucnorm}
\alias{rand_right_trucnorm}
\title{Simulate right truncated normal distribution}
\usage{
rand_right_trucnorm(n, mu, sigma, upper, ratio = 1)
}
\arguments{
\item{n}{sample size}

\item{mu}{the location parameter}

\item{sigma}{the scale parameter}

\item{upper}{the upper bound}
}
\value{
a vector of random numbers from a right truncated
normal distribution
}
\description{
Simulate right truncated normal distribution
}
\examples{
r <- rand_right_trucnorm(1000, mu=100, sigma=0.1, upper=90)
hist(r)
}
\author{
Jian Kang <jiankang@umich.edu>
}
