% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_linear_reg}
\alias{sim_linear_reg}
\title{Simulate data from the logistic regression model}
\usage{
sim_linear_reg(n, p, R2, beta_nonzero)
}
\arguments{
\item{n}{sample size}

\item{p}{number of candidate predictors}

\item{q}{number of nonzero predictors}

\item{beta_size}{effect size of beta coefficients}
}
\value{
a list objects consisting of the following components
\describe{
\item{delta}{vector of n outcome variables}
\item{X}{n x p matrix of candidate predictors}
\item{betacoef}{vector of p regression coeficients}
\item{R2}{R-squared indicating the proportion of variation explained by the predictors}
}
}
\description{
Simulate data from the logistic regression model
}
\examples{
dat <- sim_linear_reg(100,100,0.5,c(1,-1,1))
}
\author{
Jian Kang <jiankang@umich.edu>
}
