% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBayesReg.R
\name{sim_logit_reg}
\alias{sim_logit_reg}
\title{Simulate data from the logistic regression model}
\usage{
sim_logit_reg(n = 100, p = 10, q = 5, beta_size = 1)
}
\arguments{
\item{n}{sample size}

\item{p}{number of candidate predictors}

\item{q}{number of nonzero predictors}

\item{beta_size}{effect size of beta coefficients}
}
\value{
a list objects consisting of the following components
\describe{
\item{delta}{vector of n outcome variables}
\item{X}{n x p matrix of candidate predictors}
\item{betacoef}{vector of p regression coeficients}
\item{prob}{vector of n outcome probabilities}
\item{R2}{R-squared indicating the proportion of variation explained by the predictors}
}
}
\description{
Simulate data from the logistic regression model
}
\examples{
dat <- sim_logit_reg()
}
\author{
Jian Kang <jiankang@umich.edu>
}
