% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_multiclass_reg}
\alias{sim_multiclass_reg}
\title{Simulate data from the multinomial logistic regression model}
\usage{
sim_multiclass_reg(
  K = 5L,
  n = 100L,
  p = 20L,
  q = 5L,
  X_cor = 0.5,
  X_var = 10,
  beta_size = 1,
  intercept_size = 1,
  intercept0 = NULL
)
}
\arguments{
\item{K}{number of class}

\item{n}{sample size}

\item{p}{number of candidate predictors}

\item{q}{number of nonzero predictors}

\item{X_cor}{correlation between covariates}

\item{X_var}{marginal variance of covariates}

\item{beta_size}{effect size of beta coefficients}

\item{intercept_size}{intercepts}

\item{intercept0}{a vector of (K-1) intercepts}

\item{R2}{R-squared indicating the proportion of variation explained by the predictors}
}
\value{
a list objects consisting of the following components
\describe{
\item{y}{vector of n outcome variables}
\item{X}{n x p matrix of candidate predictors}
\item{betacoef}{n x (K-1) matrix of regression coeficients}
\item{intercept}{a vector of (K-1) intercept coeficients}
\item{R2}{R-squared indicating the proportion of variation explained by the predictors}
\item{sigma2}{noise variance}
\item{X_cor}{correlation between covariates}
}
}
\description{
Simulate data from the multinomial logistic regression model
}
\examples{
dat<-sim_multiclass_reg(K=5,n=1000,p=20,X_var = 10,X_cor=0.5,q=10,beta_size=1,intercept0=c(5,-5,-10,-10))
glmnet_res <- with(dat,wrap_glmnet(y,cbind(1,X),family="multinomial"))
Bayes_res <- with(dat,fast_normal_multiclass(y,cbind(1,X),num_class=length(unique(y)),burnin=5000))
glmnet_pred <- as.numeric(predict(glmnet_res$glmnet_fit,newx = cbind(1,dat$X),type = "class"))
Bayes_pred <- apply(Bayes_res$post_mean$prob,1,which.max)-1
print(c(glmnet_acc = mean(glmnet_pred==dat$y),Bayes_acc = mean(Bayes_pred==dat$y)))

}
\author{
Jian Kang <jiankang@umich.edu>
}
