% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{special_rmvnorm}
\alias{special_rmvnorm}
\title{Sample special form of multivariate normal distribution given
precision matrix and precision matrix weighted mean parameters}
\usage{
special_rmvnorm(n, mu, Omega)
}
\arguments{
\item{n}{sample size}

\item{mu}{precision weighted mean vector}

\item{Omega}{precision matrix}
}
\value{
a matrix of sample each column follows a multivariate normal distribution
}
\description{
Sample special form of multivariate normal distribution given
precision matrix and precision matrix weighted mean parameters
}
\examples{
n = 1000
d = 100
mean_X <- seq(1,d,length=d)
cov_X <- matrix(0.5,nrow=d,ncol=d) + 0.5*diag(d)
Omega = solve(cov_X)
mu = Omega\%*\%mean_X
X <- special_rmvnorm(100,mu,Omega)
}
\author{
Jian Kang <jiankang@umich.edu>
}
