% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBayesReg.R
\name{train_test_splits}
\alias{train_test_splits}
\title{Training test data split}
\usage{
train_test_splits(n, train_prop = 0.8, num_split = 1)
}
\arguments{
\item{n}{an integer number indicating the sample size of the whole dataset.}

\item{train_prop}{a number between 0 and 1 indicating the proportion of the training data}

\item{num_split}{an integer indicating the number of splits to generate}
}
\value{
a list of objects including the training and test indices.
}
\description{
Training test data split
}
\examples{
one_split <- train_test_splits(100,num_split=1)
multiple_splits <- train_test_splits(100,num_split=20)
}
\author{
Jian Kang <jiankang@umich.edu>
}
