% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_logit_reg}
\alias{sim_logit_reg}
\title{Simulate data from the logistic regression model}
\usage{
sim_logit_reg(
  n = 100L,
  p = 20L,
  q = 5L,
  X_cor = 0.5,
  X_var = 10,
  beta_size = 1,
  density = 1
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of candidate predictors}

\item{q}{number of nonzero predictors}

\item{X_cor}{correlation between covariates}

\item{beta_size}{effect size of beta coefficients}

\item{density}{specifies the percentage of non-zero elements. The default value is 1.0 in which case
X is a dense matrix. When the density is strictly less than 1.0 in which case X is a spare matrix.}

\item{R2}{R-squared indicating the proportion of variation explained by the predictors}
}
\value{
a list objects consisting of the following components
\describe{
\item{y}{vector of n outcome variables}
\item{X}{n x p matrix of candidate predictors}
\item{betacoef}{vector of p regression coeficients}
\item{R2}{R-squared indicating the proportion of variation explained by the predictors}
\item{sigma2}{noise variance}
\item{X_cor}{correlation between covariates}
}
}
\description{
Simulate data from the logistic regression model
}
\examples{
dat<-sim_logit_reg(n=1000,p=20,X_var = 1,X_cor=0.9,q=10)
summary(with(dat,glm(y~0+X,family=binomial())))
}
\author{
Jian Kang <jiankang@umich.edu>
}
